requirements.txt
BCEmbedding==0.1.5
transformers==4.45.0
streamlit==1.36.0
gradio==5.0.2
sentencepiece==0.2.0
accelerate==0.30.1
transformers_stream_generator==0.0.5
sentence-transformers==3.0.1
peft==0.11.1
xtuner==0.1.23 
openxlab
tiktoken
einops
oss2
requests
langchain==0.2.10
langchain_community==0.2.9
langchain_core
langchain-huggingface==0.0.3
langchain_text_splitters==0.2.2
chromadb==0.5.0
loguru==0.7.2
modelscope==1.18.0
numpy==1.26.4
pandas==2.2.2
timm==1.0.8 
openai==1.40.3 
lmdeploy[all]==0.5.3
hydra-core==1.3.2